Removed csr and ssr options for one-comic view
